2021/2021-07-12/dean/03-dynamodb/requirements.txt
-e .
aws_cdk.aws_lambda
aws_cdk.aws_apigateway
aws_cdk.aws_dynamodb